requirements.txt
# Core dependencies for semantic media search
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
opencv-python>=4.8.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
Pillow>=10.0.0

python-multipart>=0.0.6  # For FastAPI file uploads if needed
ruff>=0.1.0 # For code formatting and linting
